refactor: Extract fetch hook
